+++ packages/orq-rc/RELEASES.md
@@ -3899,3 +3899,13 @@ Based on:
 - [typescript v3.13.0-rc.42] packages/orq-rc
 ### Releases
 - [NPM v3.13.0-rc.42] https://www.npmjs.com/package/@orq-ai/node/v/3.13.0-rc.42 - packages/orq-rc
+
+## 2025-09-29 07:37:26
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.631.4 (2.716.16) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v3.13.0-rc.43] packages/orq-rc
+### Releases
+- [NPM v3.13.0-rc.43] https://www.npmjs.com/package/@orq-ai/node/v/3.13.0-rc.43 - packages/orq-rc